\name{algo.call}
\alias{algo.call}
\title{Query Transmission to Specified Surveillance Systems}
\description{Transmission of a object of class disProg to the specified surveillance systems.
}
\usage{
    algo.call(disProgObj, control = list(
                     list(funcName = "rki1", range = range),
                     list(funcName = "rki", range = range,
                          b = 2, w = 4, actY = TRUE),
                     list(funcName = "rki", range = range,
                          b = 2, w = 5, actY = TRUE)))
}
\arguments{
        \item{disProgObj}{object of class disProg, which includes the state chain and the observed.}
        \item{control}{specifies which surveillance systems should be used with their parameters.
                The parameter \code{funcName} and \code{range} must be specified where funcName must be
                the apropriate method function (without 'algo.'). \code{range} defines
                the timepoints to be evaluated by the actual system. If \code{control} includes \code{name}
                this name is used in the survRes Object as name.}
}
\value{
        \item{list of survRes Objects}{generated by the specified surveillance systems.
        }
}
\seealso{\code{\link{algo.rki}}, \code{\link{algo.bayes}}, \code{\link{algo.farrington}}}
\examples{
    # Create a test object
    disProgObj <- sim.pointSource(p = 0.99, r = 0.5, length = 400, A = 1,
                                    alpha = 1, beta = 0, phi = 0,
                                    frequency = 1, state = NULL, K = 1.7)

    # Let this object be tested from any methods in range = 200:400
    range <- 200:400
    survRes <- algo.call(disProgObj,
                  control = list(
                            list(funcName = "rki1", range = range),
                            list(funcName = "rki2", range = range),
                            list(funcName = "rki3", range = range),
                            list(funcName = "rki", range = range,
                                 b = 3, w = 2, actY = FALSE),
                            list(funcName = "rki", range = range,
                                 b = 2, w = 9, actY = TRUE),
                            list(funcName = "bayes1", range = range),
                            list(funcName = "bayes2", range = range),
                            list(funcName = "bayes3", range = range),
                            list(funcName = "bayes", name = "myBayes",
                                 range = range, b = 1, w = 5, actY = TRUE,alpha=0.05)
                             ) )
    # this are some survResObjects
    survRes$rki1
    survRes$myBayes
}

\keyword{}


