\name{algo.farrington.assign.weights}
\alias{algo.farrington.assign.weights}
\encoding{latin1}

\title{Assign weights to base counts}
\description{
  Weights are assigned according to the Anscombe residuals
}
\usage{
algo.farrington.assign.weights(s)
}
\arguments{
  \item{s}{Vector of standardized Anscombe residuals}
}
\value{Weights according to the residuals}
\seealso{See Also as \code{\link{anscombe.residuals}}}
\keyword{regression}


