\name{CIdata}
\alias{CIdata}
\title{Confidence-Interval for the Mean of the Poisson Distribution}
\description{In the first column the mean from 0 to 20 is shown,
	In the second the lower and in the third the upper value of
	the 95 percent confidence interval. These intervals are used in
	the RKI Algorithms.
}
\usage{data(CIdata)}
\format{A data frame with header.}
\source{L. Sachs. Angewandte Statistik. Springer Verlag, 7. Auflage, S.446, 1991}
\examples{
    require(surveillance)
    data(CIdata)
}
\seealso{\code{\link{algo.rki}}}
\keyword{data}

