\name{hepatitisA}
\alias{hepatitisA}
\title{Hepatitis A in Germany}
\description{
  Missing!!!!
}
\usage{data(hepatitisA)
}
\format{
  A \code{disProg} object containing \eqn{208\times 1}{208 x 1}
  observations starting from week 1 in 2001 to week 30 in 2006. 
}
\source{
Robert Koch-Institut: SurvStat: \url{http://www3.rki.de/SurvStat};
Queried on XX YY 2006.

}
\examples{
data(hepatitisA)
plot(hepatitisA)
}
\keyword{datasets}


