\name{marks}
\alias{marks}
\docType{import}
\title{Import from package \pkg{spatstat.geom}}
\description{
The generic function \code{marks} is imported from package \pkg{spatstat.geom}.
See \code{\link[spatstat.geom:marks]{spatstat.geom::marks}} for \pkg{spatstat.geom}'s
own methods, and \code{\link{marks.epidataCS}} for the
\code{"epidataCS"}-specific method.
}
