\name{powerlaw}
\alias{powerlaw}
\title{
  Power-Law Neighbourhood Weight Structure for \code{hhh4} Models
}

\description{
  Set up a power-law weight structure for the neighbourhood
  component of \code{\link{hhh4}} models -- STILL EXPERIMENTAL.
}

% \usage{
% powerlaw(maxlag, normalize = TRUE, initial = 1)
% }

% \arguments{
%   \item{maxlag}{}
%   \item{normalize}{}
%   \item{initial}{}
%   \item{nbmat}{}
%   \item{d}{}
%   \item{shared}{}
% }

%\value{}

%\references{}

\author{
  Sebastian Meyer
}

%\seealso{}

%\examples{}

\keyword{spatial}
\keyword{models}
\keyword{utilities}
