\name{ranef}
\alias{ranef}
\alias{fixef}
\docType{import}
\title{Import from package \pkg{nlme}}
\description{
  The generic functions \code{ranef} and \code{fixef}
  are imported from package \pkg{nlme}.
  See \code{\link[nlme:ranef]{nlme::ranef}} for \pkg{nlme}'s own
  description, and \code{\link{ranef.hhh4}} or \code{\link{fixef.hhh4}}
  for the added methods for \code{"\link{hhh4}"} models.
}
