\encoding{latin1}
\name{residualsCT}
\alias{residuals.twinSIR}
\alias{residuals.twinstim}
\title{
~~ not documented yet ... ~~
}
\description{
~~ not documented yet ... ~~
}
\usage{
\method{residuals}{twinSIR}(object, plot = TRUE, ...)
\method{residuals}{twinstim}(object, plot = TRUE, ...)
}
\arguments{
  \item{object}{
    an object of class \code{"\link{twinSIR}"} or \code{"\link{twinstim}"}.
}
  \item{plot}{
    logical indicating if a plot of the transformed residuals should be
    produced by \code{\link{ks.plot.unif}} to check for deviations
    from the uniform distribution.
}
  \item{\dots}{
    further arguments passed to \code{\link{ks.plot.unif}}.
}
}
\value{
~~ not documented yet ... ~~
}
\references{
~~ not documented yet ... ~~
}
\author{
Michael H\enc{ö}{oe}hle and Sebastian Meyer
}
\examples{
#TODO
}
\keyword{aplot}
\keyword{htest}

