\name{rotaBB}
\alias{rotaBB}
\docType{data}
\title{Rotavirus cases in Germany 2002-2013 in 5 age categories}
\description{
  Reported number of cases of infections with rotavirus in 5 age categories (00-04, 05-09, 10-14, 15-69, 70+) 2002-2013.
}
\usage{data(rotaBB)}
\format{
  A \code{sts} object. 
}
\source{
  The data are queried from the Survstat@RKI database of the German
  Robert Koch Institute (http://www3.rki.de/SurvStat/). 
}  
\examples{
data(rotaBB)
}
\keyword{datasets}
