\docType{data}
\name{salmAllOnset}
\alias{salmAllOnset}
\title{Salmonella cases in Germany 2001-2014 by data of symptoms onset}
\format{A sts-object}
\usage{
data(salmAllOnset)
}
\description{
A dataset containing the reported number of cases of Salmonella in Germany 2001-2014 aggregated by data of disease onset. The slot \code{control} contains a matrix \code{reportingTriangle$n} with the reporting triangle as described in Salmon et al.
}
\references{
Bayesian outbreak detection in the presence of reporting delays, Salmon, M., Schumacher, D., Stark, K., \enc{H?hle}{Hoehle}, M. (2015), in revision.
}
\keyword{datasets}
