\name{salmNewport}
\alias{salmNewport}
\docType{data}
\title{Salmonella Newport cases in Germany 2004-2013}
\description{
  Reported number of cases of the Salmonella Newport serovar in German federal states
   2004-2013. Survstat.
}
\usage{data(salmNewport)}
\format{
  A \code{sts} object. 
}
\source{An outbreak of
Salmonella
Newport associated with
mung bean sprouts in Germany and the Netherlands,
October to November 2011, Bayer, C., Bernard, H., Prager, R., Rabsch, W., Hiller, P., Malorny, B., Pfefferkorn, B., Frank, C., de Jong, A., Friesema, I., Start, K., Rosner, B.M., Eurosurveillance, 2014.}
\examples{
data(salmNewport)
}
\keyword{datasets}
