\name{stsNewport}
\alias{stsNewport}
\docType{data}
\title{Salmonella Newport cases in Germany 2004-2013}
\description{
  Reported number of cases of the Salmonella Newport serovar in the 16 German
  federal states 2001-2015, by date of disease onset. The slot \code{control} contains a matrix \code{reportingTriangle$n} with the reporting triangle as described in Salmon et al.
}
\usage{data(stsNewport)}
\format{
  A \code{sts} object. 
}


\references{
Bayesian outbreak detection in the presence of reporting delays, Salmon, M., Schumacher, D., Stark, K., \enc{H?hle}{Hoehle}, M. (2015), in revision.
}
\examples{
  data(stsNewport)
}
\keyword{datasets}