\name{year-methods}
\docType{methods}
\alias{year}
\alias{year-methods}
\alias{year,sts-method}
\title{ Methods for Function year  }
\description{
 Method to extract the corresponding year of each obervatuib in an
 \code{sts} object. The year is calculated based on the \code{start} slot.
}
\section{Methods}{
\describe{

\item{x = "sts"}{ For each time point in the \code{x} the corresponding
  year is determined and returned. }
}}
\keyword{methods}
