\name{surveillance-defunct}
\alias{surveillance-defunct}
\title{Defunct Functions in Package \pkg{surveillance}}
\alias{testSim}
\description{
  The functions listed here are no longer part of \pkg{surveillance}.
}
\usage{
## Defunct in surveillance 1.17.0
testSim(p = 0.99, r = 0.01, length = 400, A = 1, alpha = 1,
        beta = 0, phi = 0, frequency = 1, state = NULL, K, 
        range = 200:400)
}
\details{
  The ancient test function \code{testSim} is now provided as an example
  in \code{help("\link{sim.pointSource}")}.
}
\seealso{
  \code{\link{Defunct}}
}
\keyword{internal}
