\name{zetaweights}
\alias{zetaweights}
\title{
  Derive Power-Law Weights from Neighbourhood Order
}

\description{
  STILL EXPERIMENTAL...
}

% \usage{
% zetaweights(nbmat, maxlag = Inf, rho = 1, normalize = TRUE, shared = FALSE)
% }

% \arguments{
%   \item{nbmat}{}
%   \item{maxlag}{}
%   \item{rho}{}
%   \item{normalize}{}
%   \item{shared}{}
% }

%\value{}

%\references{}

\author{
  Sebastian Meyer
}

%\seealso{}

%\examples{}

\keyword{spatial}
\keyword{utilities}
