\name{animate}
\alias{animate}

\title{
Generic animation of spatio-temporal objects
}

\description{
Generic function for animation of \R objects.
}

\usage{
animate(object, ...)
}

\arguments{
  \item{object}{The object to animate.}
  \item{\dots}{
    Arguments to be passed to methods, such as graphical parameters or
    time interval options for the snapshots.
  }
}

\seealso{
The methods \code{\link{animate.epidata}}, \code{\link{animate.epidataCS}},
and \code{\link{animate.sts}} for the animation of surveillance data.
}

\keyword{hplot}
\keyword{dynamic}
\keyword{spatial}
