\name{hepatitisA}
\docType{data}
\alias{hepatitisA}
\title{Hepatitis A in Germany}
\description{
  Weekly number of reported hepatitis A infections in Germany 2001-2004. 
}
\usage{data(hepatitisA)}
\format{
  A \code{disProg} object containing \eqn{208\times 1}{208 x 1}
  observations starting from week 1 in 2001 to week 52 in 2004. 
}
\source{
Robert Koch-Institut: SurvStat: \url{https://survstat.rki.de/};
Queried on 11-01-2005.
}
\examples{
data(hepatitisA)
plot(hepatitisA)
}
\keyword{datasets}
