%This is just a way to "hide" functions, which the user should not know
%the proper way is to use NAMESPACE here (TODO!)
\name{surveillance-internal}
\alias{anscombeNB}
\alias{hypgeom2F1special}
\alias{hypGeomSmall}
%Some S4 stuff from sts
\alias{add.legend}
\alias{fix.dimnames}
\alias{hcl.colors}
\alias{insert.zeroes}
\alias{wait}
\alias{init.sts}
%Functions from M.Paul
\alias{commonDenom}
\alias{logInterpolation}
\alias{getH}
\alias{getHK}
\alias{getTheta1}
%
\title{Internal surveillance Functions}
\description{
  Undocumented internal \code{surveillance} functions. If you really
  want to know whats going on you might consider looking at the source
  code. If the package had a namespace you would not see this list. 
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
