\name{magic.dim}
\alias{magic.dim}
\encoding{latin1}

\title{Returns a suitable k1 x k2 for plotting the disProgObj}
\description{
       For a given number k \code{magic.dim} provides a vector 
       containing two elements nRows and nCols which can be used to set the 
       dimension of a single graphic device so that nRow*nCol plots can be 
       drawn by row (or by column) on the device.
}
\usage{
      magic.dim(k)
}
\arguments{
      \item{k}{an integer}
}
\value{vector with two elements}
\keyword{hplot}


% primeFactors
