\name{multiplicity}
\alias{multiplicity}
\docType{import}
\title{Import from package \pkg{spatstat.geom}}
\description{
The generic function \code{multiplicity} is imported from package \pkg{spatstat.geom}.
See \code{\link[spatstat.geom:multiplicity]{spatstat.geom::multiplicity}} for
\pkg{spatstat.geom}'s own methods, and \code{\link{multiplicity.Spatial}} for the
added method for \code{"\linkSPclass{Spatial}"} objects.
}
