\name{multiplicity}
\alias{multiplicity}
\docType{import}
\title{Import from package \pkg{spatstat}}
\description{
The generic function \code{multiplicity} is imported from package \pkg{spatstat}.
See \code{\link[spatstat:multiplicity]{spatstat::multiplicity}} for
\pkg{spatstat}'s own methods, and \code{\link{multiplicity.Spatial}} for the
added method for \code{\linkS4class{Spatial}} objects.
}
