\name{primeFactors}
\alias{primeFactors}
\encoding{latin1}

\title{Prime number factorization}
\description{
       Computes prime number factorization of an integer x. 
}
\usage{
      primeFactors(x)
}
\arguments{
      \item{x}{an integer}
}
\value{vector with prime number factorization of x}
%\keyword{}


% bestCombination
