\name{rotaBB}
\alias{rotaBB}
\docType{data}
\title{Rotavirus cases in Brandenburg, Germany, during 2002-2013 stratified by 5 age categories}
\description{
  Monthly reported number of rotavirus infections in the federal state of
  Brandenburg stratified by five age categories (00-04, 05-09, 10-14, 15-69, 70+)
  during 2002-2013.
}
\usage{data(rotaBB)}
\format{
  A \code{sts} object. 
}
\source{
  The data were queried on 19 Feb 2014 from the Survstat@RKI database of the German
  Robert Koch Institute (\url{https://survstat.rki.de/}). 
}  
\keyword{datasets}
