\docType{data}
\name{salmAllOnset}
\alias{salmAllOnset}
\title{Salmonella cases in Germany 2004-2014 by data of symptoms onset}
\format{A sts-object}
\usage{
data(salmAllOnset)
}
\description{
A dataset containing the reported number of cases of Salmonella in Germany 2004-2014 aggregated by data of disease onset
}
\references{
The data are queried from the Survstat@RKI database of the German Robert Koch Institute (http://www3.rki.de/SurvStat/).
}
\keyword{datasets}
