\name{salmHospitalized}
\alias{salmHospitalized}
\docType{data}
\title{Hospitalized Salmonella cases in Germany 2004-2014}
\description{
  Reported number of cases of Salmonella in Germany 2004-2014 (early 2014) that were hospitalized. The corresponding 
  total number of cases is indicated in the slot \code{populationFrac} and \code{multinomialTS} is \code{TRUE}.
}
\usage{data(salmHospitalized)}
\format{
  An \code{"\linkS4class{sts}"} object. 
}
\source{
  The data are queried from the Survstat@RKI database of the German
  Robert Koch Institute (\url{https://survstat.rki.de/}).
}
\keyword{datasets}
