\name{salmHospitalized}
\alias{salmHospitalized}
\docType{data}
\title{Hospitalized Salmonella cases in Germany 2004-2013}
\description{
  Reported number of cases of Salmonella in Germany 2004-2013 that were hospitalized. The corresponding 
  total number of cases is indicated in the \code{slot populationFrac}. \code{multinomialTS} is \code{TRUE}.
}
\usage{data(salmNewport)}
\format{
  A \code{sts} object. 
}
%\source{}
\examples{
data(salmHospitalized)
}
\keyword{datasets}
