\name{salmNewport}
\alias{salmNewport}
\docType{data}
\title{Salmonella Newport cases in Germany 2004-2013}
\description{
  Reported number of cases of the Salmonella Newport serovar in the 16 German
  federal states 2004-2013.
}
\usage{data(salmNewport)}
\format{
  A \code{sts} object. 
}
\source{
  The data were queried from the SurvStat@RKI database of the German
  Robert Koch Institute (\url{https://survstat.rki.de/}). A detailed
  description of the 2011 outbreak can be found in the publication

  Bayer, C., Bernard, H., Prager, R., Rabsch, W., Hiller, P., Malorny,
  B., Pfefferkorn, B., Frank, C., de Jong, A., Friesema, I., Start, K.,
  Rosner, B.M. (2014), An outbreak of Salmonella Newport associated with
  mung bean sprouts in Germany and the Netherlands, October to November
  2011, Eurosurveillance 19(1):pii=20665.
}  
\keyword{datasets}
