\name{salmonella.agona}
\alias{salmonella.agona}
\docType{data}
\title{Salmonella Agona cases in the UK 1990-1995}
\description{
  Reported number of cases of the Salmonella Agona serovar in the UK
  1990-1995. Note however
  that the counts do not correspond exactly to the ones used by
  Farrington et. al (1996).
}
\usage{data(salmonella.agona)}
\format{
  A \code{disProg} object with 312 observations starting from week 1 in 1990.
}
\source{
  A statistical algorithm for the early detection of outbreaks of
  infectious disease, Farrington, C.P., Andrews, N.J, Beale A.D. and
  Catchpole, M.A. (1996). , J. R. Statist. Soc. A, 159, 547-563.
}
\keyword{datasets}
