\encoding{latin1}
\name{stsNewport}
\alias{stsNewport}
\docType{data}
\title{Salmonella Newport cases in Germany 2004-2013}
\description{
  Reported number of cases of the Salmonella Newport serovar in Germany
  2001-2015, by date of disease onset. The slot \code{control} contains
  a matrix \code{reportingTriangle$n} with the reporting triangle as
  described in Salmon et al. (2015).
}
\usage{data(stsNewport)}
\format{
  A \code{sts} object. 
}

\references{
  Salmon, M., Schumacher, D., Stark, K., \enc{Höhle}{Hoehle}, M. (2015):
  Bayesian outbreak detection in the presence of reporting delays.
  Biometrical Journal, 57 (6), 1051-1067.
}
\keyword{datasets}
