\name{observed-methods}
\docType{methods}
\alias{observed}
\alias{observed<-}
\alias{observed-methods}
\alias{alarms}
\alias{alarms<-}
\alias{upperbound}
\alias{upperbound<-}
\alias{alarms-methods}
\alias{control}
\alias{control<-}
\alias{control-methods}
\alias{epoch}
\alias{epoch<-}
\alias{epoch-methods}
\alias{epochInYear}
\alias{epochInYear<-}
\alias{epochInYear-methods}
\alias{population}
\alias{population<-}
\alias{population-methods}
\title{ Methods }
\description{
 Method to extract or set the corresponding slot of an
 \code{sts} object. This documentation is not really valid yet.
}
\section{Methods}{
\describe{

\item{x = "sts"}{ The slot of \code{x} is determined and
  returned or set.}
}}

\keyword{methods}
