\name{stsSlot-generics}
\docType{methods}
\alias{alarms}
\alias{alarms<-}
\alias{upperbound}
\alias{upperbound<-}
\alias{control}
\alias{control<-}
\alias{epoch}
\alias{epoch<-}
\alias{observed}
\alias{observed<-}
\alias{population}
\alias{population<-}
\alias{multinomialTS}
\alias{multinomialTS<-}
\alias{neighbourhood}
\alias{neighbourhood<-}

\title{Generic Functions to Access \code{"sts"} Slots}
\description{
 For almost every slot of the \code{"sts"} class, package \pkg{surveillance}
 defines a generic function of the same name (and a replacement
 version) to extract (or set) the corresponding slot.
 See the \code{"\linkS4class{sts}"} class documentation.
}
\keyword{methods}
