\name{surveillance-deprecated}
\title{Deprecated Functions in Package \pkg{surveillance}}
\alias{surveillance-deprecated}
\alias{qlomax}
\description{
  The functions listed here are provided for compatibility
  with older versions of \pkg{surveillance} only,
  and may be defunct as soon as of the next release.
}
\usage{
qlomax(p, scale, shape)
}
\arguments{
  \item{p}{vector of probabilities.}
  \item{scale}{positive scale parameter.}
  \item{shape}{positive shape parameter.}
}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
