\encoding{latin1}
\name{surveillance-package}
\alias{surveillance-package}
\alias{surveillance}
\docType{package}
\title{
  Outbreak detection algorithms for surveillance data
}
\description{
  surveillance is an \R package for the detection of abberations
  in routine collected public health data seen as univariate and
  multivariate time series of counts.

  As many methods rely on statistical process control methodology the
  package also contains an implementation of count data time series
  methods relevant to quality control and reliability engineering.
} 

\details{
  \tabular{ll}{
    Package: \tab surveillance\cr
    Type: \tab Package\cr
    Version: \tab 1.0-1\cr
    Date: \tab 2009-05-03\cr
    License: \tab GPL version 2 (http://www.gnu.org/licenses/gpl.html) \cr
  }
  \itemize{
    %~~ An overview of how to use the package, including the most important functions ~~
  }
}
\author{
  Author: M. \enc{Höhle}{Hoehle} with contributions from T. Correa, M. Hofmann,
  C. Lang, M. Paul, A. Riebler and V. Wimmer
  
  Maintainer: Michael \enc{Höhle}{Hoehle} \email{<hoehle@stat.uni-muenchen.de>}
}
\references{
  surveillance: An R package for the surveillance of infectious diseases (2007), M. \enc{Höhle}{Hoehle}, Computational Statistics, 22(4), pp. 571---582.
}
\keyword{ package }
\examples{
\dontrun{demo(cost)}
}
