\name{twinSIR_exData}
\alias{fooepidata}
\alias{foofit}
\docType{data}

\title{
Toy Data for \code{twinSIR}
}

\description{
  This is data, which is used in \code{twinSIR} examples.
  
  \code{fooepidata} is an object of class \code{"\link{epidata}"}.
  \code{foofit} is an object of class
  \code{"\link{twinSIR}"} and is the result of a fit to \code{fooepidata}.
}

\usage{
data(fooepidata)
data(foofit)
}

\keyword{datasets}
\keyword{internal}
