\name{twinSIR_exData}
\alias{fooepidata}
\docType{data}

\title{
Toy Data for \code{twinSIR}
}

\description{
  Toy \code{"\link{epidata}"} previously used to exemplify
  \code{\link{twinSIR}} models. We now use the well-known
  \code{\link{hagelloch}} dataset.
}

\usage{
data(fooepidata)
}

\keyword{datasets}
\keyword{internal}
